fix bug of nbt_doPLI.m
The instantaneous phase computed with angle(hilbert(Signal)) is between
-pi and pi  , therefore, the phase difference range from -2pi to 2pi ,
but the Phase Lag Index is defined on a phase difference in [-pi,pi]. So
the phase difference must be project to [-pi , pi  ]  or compute the PLI
use  abs(mean(sign(sin(phaseSignal(:,i)-phaseSignal(:,m)))).   (Stam.
et.al , 2007)